# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeTextBody {
  /** The text or content to analyze */
  text: string;
}

/**
 * Scores (0-100) for each detected manipulation technique
 */
export interface ManipulationBreakdown {
  /** Fear tactics score (0-100) */
  fear: number;
  /** Urgency language score (0-100) */
  urgency: number;
  /** Emotional trigger score (0-100) */
  emotionalTriggers: number;
  /** Fake authority signals score (0-100) */
  fakeAuthority: number;
}

export type AnalysisResultRiskLevel =
  (typeof AnalysisResultRiskLevel)[keyof typeof AnalysisResultRiskLevel];

export const AnalysisResultRiskLevel = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface AnalysisResult {
  /** Score from 0 to 100 (100 = fully credible) */
  credibilityScore: number;
  riskLevel: AnalysisResultRiskLevel;
  /** Human-readable explanation of the analysis */
  explanation: string;
  /** List of suspicious phrases found in the text */
  suspiciousPhrases: string[];
  manipulationBreakdown: ManipulationBreakdown;
  /** Record ID for history tracking */
  id: string;
}

export type AnalysisRecordRiskLevel =
  (typeof AnalysisRecordRiskLevel)[keyof typeof AnalysisRecordRiskLevel];

export const AnalysisRecordRiskLevel = {
  Low: "Low",
  Medium: "Medium",
  High: "High",
} as const;

export interface AnalysisRecord {
  id: string;
  /** First 200 chars of analyzed text */
  textSnippet: string;
  credibilityScore: number;
  riskLevel: AnalysisRecordRiskLevel;
  explanation: string;
  suspiciousPhrases: string[];
  manipulationBreakdown: ManipulationBreakdown;
  analyzedAt: string;
}

export type AnalysisStatsRiskDistribution = {
  Low: number;
  Medium: number;
  High: number;
};

export interface AnalysisStats {
  totalAnalyses: number;
  avgCredibilityScore: number;
  riskDistribution: AnalysisStatsRiskDistribution;
}