# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analysis/analyze:
    post:
      operationId: analyzeText
      tags: [analysis]
      summary: Analyze text for misinformation
      description: Analyze input text for credibility, risk level, suspicious phrases, and manipulation technique breakdown
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeTextBody"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalysisResult"
...
components:
  schemas:
...
    AnalyzeTextBody:
      type: object
      properties:
        text:
          type: string
          description: The text or content to analyze
      required:
        - text
    ManipulationBreakdown:
      type: object
      description: Scores (0-100) for each detected manipulation technique
      properties:
        fear:
          type: number
          description: Fear tactics score (0-100)
        urgency:
          type: number
          description: Urgency language score (0-100)
        emotionalTriggers:
          type: number
          description: Emotional trigger score (0-100)
        fakeAuthority:
          type: number
          description: Fake authority signals score (0-100)
      required:
        - fear
        - urgency
        - emotionalTriggers
        - fakeAuthority
    AnalysisResult:
      type: object
      properties:
        credibilityScore:
          type: number
          description: Score from 0 to 100 (100 = fully credible)
        riskLevel:
          type: string
          enum: [Low, Medium, High]
        explanation:
          type: string
          description: Human-readable explanation of the analysis
        suspiciousPhrases:
          type: array
          items:
            type: string
          description: List of suspicious phrases found in the text
        manipulationBreakdown:
          $ref: "#/components/schemas/ManipulationBreakdown"
        id:
          type: string
          description: Record ID for history tracking
      required:
        - credibilityScore
        - riskLevel
        - explanation
        - suspiciousPhrases
        - manipulationBreakdown
        - id
...