# lib/api-zod/src/generated/types/analysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AnalysisResultRiskLevel } from "./analysisResultRiskLevel";
import type { ManipulationBreakdown } from "./manipulationBreakdown";

export interface AnalysisResult {
  /** Score from 0 to 100 (100 = fully credible) */
  credibilityScore: number;
  riskLevel: AnalysisResultRiskLevel;
  /** Human-readable explanation of the analysis */
  explanation: string;
  /** List of suspicious phrases found in the text */
  suspiciousPhrases: string[];
  manipulationBreakdown: ManipulationBreakdown;
  /** Record ID for history tracking */
  id: string;
}