# lib/api-zod/src/generated/types/manipulationBreakdown.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Scores (0-100) for each detected manipulation technique
 */
export interface ManipulationBreakdown {
  /** Fear tactics score (0-100) */
  fear: number;
  /** Urgency language score (0-100) */
  urgency: number;
  /** Emotional trigger score (0-100) */
  emotionalTriggers: number;
  /** Fake authority signals score (0-100) */
  fakeAuthority: number;
}